requirements.txt
aiofiles==0.3.1
httptools==0.0.9
Jinja2==2.9.6
MarkupSafe==1.0
python-engineio==1.5.1
python-socketio==1.7.4
sanic==0.5.4
sanic-jinja2==0.5.1
sh==1.12.13
six==1.10.0
ujson==1.35
uvloop==0.8.0
websockets==3.3
